Pass a single argument to actions
If we ever use this functionality, it will be a single string
coming in that will have to be parsed.  If that ever becomes
cumbersome we can insert a parser, but frankly it looks
like these functions do not really even need an any arguments.
Future varpor ware should not influence design too much