Revert "Always use f->win as focus when f->p is null"
This reverts commit 4ce1eebf3402e6e73843d536e57cb10081143716.

We need to have nodelay set or else the wget_wch in the main
loop does not return.  We are currently deleting the window
if h or w is 0, but we want to be able to focused the
hidden window, so we highlight the divider.  So...either we
restructure a lot (probably should), or we just set nodelay
on the titles and dividers.