Adjust scroll region in ris
I am not really sure what this should be.  Clearly, we should be
subtracting one, but I do not understand how the scroll region is
supposed to work off the top of the screen.  It seems that immediately
after setupevents() we call reshape and set the scroll region to
scrollback - 1, so any errors that may be cropping up by this
apparent off-by-one error are being masked.  Or, if this change
introduces an error, it will be masked!