Revert "Remove wait_child()"
This reverts commit 1ebff49c7ebadf80659e1dec3d28c01d88639f7b.

At the moment, I'm not sure we'll ever want to support sending
the master pty through a unix socket.  It seems easier to just
use dtach to dup the whole session, or run this in tmux as a pane
manager.  In any case, keep this for now to avoid zombies.