Check maxfd to quit
I am still undecided about how to prune canvasses.  We may want
a pool that can be used when creating, and new canvas may want
to grab from the pool of unused ptys.  But probably need to implement
a way to list those pools before we really think about this too much.